requirements.txt
numpy==1.22.4
opencv-python==4.6.0.66
pandas==1.4.3
protobuf==4.21.2
scikit-image==0.19.3
scikit-learn==1.1.1
joblib
dlib @ file:///C:/Users/elder/Downloads/dlib-19.22.99-cp310-cp310-win_amd64.whl
face-recognition @ file:///C:/Users/elder/PycharmProjects/lib_face_recognition/face_recog_dlib_file/dlib-19.19.0-cp38-cp38-win_amd64.whl
face-recognition-models==0.3.0
Flask==2.1.2
Pillow==9.2.0
tensorflow==2.9.1
tensorflow-tensorboard==0.4.0rc2
ffmpeg-python==0.1.6